Adding class material from CoderDojoGirls teaching sessions
We thought 3 streams: basic beginnersHtml, flowerGame and pianoGame. Teaching notes are
missing in a lot of cases and as two of us were swapping teaching the sessions the changeover
from week to week was sometimes inconsistent. Ideally plan to clean up this teaching material so it is more useful for others.